requirements.txt
aiohttp
aiohttp-prometheus-client
elasticsearch
gunicorn
motor
